Fin Assimp, début collisions

+++ TP4/include/Mesh.hpp
@@ -19,13 +19,13 @@ class Mesh {
     public:
         // Public attributes
         std::vector<Vertex> vertices;
-        std::vector<unsigned short> indices;
+        std::vector<unsigned int> indices;
         std::vector<Texture> textures;
         Shader shader;
 
         // Constructors
         Mesh();
-        Mesh(std::vector<Vertex> vertices, std::vector<unsigned short> indices, std::vector<Texture> textures);
+        Mesh(std::vector<Vertex> vertices, std::vector<unsigned int> indices, std::vector<Texture> textures);
 
         void bind_shader(Shader shader);
         void bind_shader(const GLchar* vertex_path, const GLchar* fragment_path);
